+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - fix(networking): allow multiple domain names in dhcp options set
+- fix(config): support port ranges in network firewall rule variable portSets
 
 ## [1.16.0] -